#179 Solved image reference

+++ app/commands/cdn.php
@@ -11,14 +11,14 @@ class cdn extends Command {
 	 *
 	 * @var string
 	 */
-	protected $name = 'cdn:publish';
+	protected $name = 'cdn';
 
 	/**
 	 * The console command description.
 	 *
 	 * @var string
 	 */
-	protected $description = 'Publish to CDN';
+	protected $description = 'CDN Commands';
 
 	/**
 	 * Create a new command instance.
@@ -37,7 +37,12 @@ class cdn extends Command {
 	 */
 	public function fire()
 	{
-		$this->info('Starting cdn:publish');
+		$action = $this->argument('action');
+
+		$publish = $action == 'publish';
+		$delete = ! $publish;
+
+		$this->info('Starting cdn ' . $action);
 		// Get all files from the CDN, store their names.  We'll be deleting those at the end.
 		
 		// Get our secret key
@@ -56,7 +61,9 @@ class cdn extends Command {
 		foreach ($objects as $object)
 			$existing_files[] = $object->getName();
 
-		$this->info(count($existing_files) . ' files found');
+		// Only show this if we're publishing
+		if ($publish)
+			$this->info(count($existing_files) . ' files found');
 
 		// Recursively go through JS and CSS.
 		// Get the md5 of the contents.  
@@ -97,22 +104,29 @@ class cdn extends Command {
 					'path' => $original_filename
 				);
 
-				$this->info('Uploading ' . $original_filename . ' as ' . $filename);
+				// Only show this if we're publishing
+				if ($publish)
+					$this->info('Uploading ' . $original_filename . ' as ' . $filename);
 			}
 
 		// Upload those files
-		$container->uploadObjects($files_to_upload);
+		// Only if we're publishing
+		if ($publish)
+			$container->uploadObjects($files_to_upload);
 
 		// Delete the necessary files
 		// if they were safe to keep, they would have been removed from this array
-		foreach ($existing_files as $ef)
-		{
-			$this->info('Deleting ' . $ef . ' from cdn');
-			$object = $container->getObject($ef);
-			$object->delete();
-		}
+		// Only delete if the command actions dictates
+		if ($delete)
+			foreach ($existing_files as $ef)
+			{
+				$this->info('Deleting ' . $ef . ' from cdn');
+				$object = $container->getObject($ef);
+				$object->delete();
+			}
 
-		$this->info('cdn:publish Finished');
+		$this->info('cdn ' . $action . ' finished');
+		$this->comment('You may want to clear cache! (php artisan cache:clear)');
 	}
 
 	/**
@@ -122,7 +136,9 @@ class cdn extends Command {
 	 */
 	protected function getArguments()
 	{
-		return [];
+		return array(
+			array('action', InputArgument::REQUIRED, 'Publish or Delete?')
+		);
 	}
 
 	/**
